requirements.txt
drf_yasg==1.20.0
Django==3.2.4
django-cors-headers==3.14.0
django-filter==22.1
django-oauth-toolkit==2.2.0
django-tenants==3.4.8
djangorestframework==3.12.4
psycopg2-binary==2.9.5
python-dotenv==0.17.1
oauthlib~=3.2.2
sendgrid~=6.9.7
Faker==18.3.1
django-debug-toolbar~=3.8.1
uwsgi>=2.0.18,<2.1